requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy

# Hugging Face / Pi0
transformers>=4.38.0
accelerate>=0.27.0
protobuf

# Utilities
psutil
tqdm
omegaconf
hydra-core
fastapi
uvicorn
sqlalchemy
python-dotenv
requests
opencv-python
scipy
pyserial
aiohttp
python-multipart
websockets
jinja2
pytest
httpx
pytest-asyncio
openmim
